transferred to NpcHanlder Chromie structure
