+++ WowPacketParserModule.V9_0_1_36216/Parsers/NpcHandler.cs
@@ -85,5 +85,18 @@ namespace WowPacketParserModule.V9_0_1_36216.Parsers
                 }
             }
         }
+		
+		[Parser(Opcode.SMSG_CHROMIE_TIME_OPEN_NPC)]
+        public static void HandleChromieTimeOpenNpc(Packet packet)
+        {
+            packet.ReadPackedGuid128("GUID");
+        }
+
+        [Parser(Opcode.CMSG_CHROMIE_TIME_SELECT_EXPANSION)]
+        public static void HandleChromieTimeSelectExpansion(Packet packet)
+        {
+            packet.ReadPackedGuid128("GUID");
+            packet.ReadUInt32("Expansion");
+        }
     }
 }
